Work around Dropbox API bug and use promises
Instead of calling callbacks from the promises the Dropbox SDK returns,
the app now uses promises for everything.
There is a bug in the Dropbox SDK where it throws an error
when trying to download a non-existent file.
I worked around that error by manually checking if the file exists
before attempting to download.