Add 'userResponse' field to modals in state
The reason for thi si sthe following:
When the user responds to a modal, the action will set the 'userResponse' field in the current modal to 'confirm' or 'cancel'.
A a saga will trigger on 'confirmModal' and 'cancelModal' to dispatch the appropriate action.
If the modal's actions were just called from the reducer, no action would actually be dispatched to the store, and redux-saga couldn't trigger, and the app wouldn't initiate a sync.