Make UI interactions trigger sync action
So far, logging a value in a metric (and other actions) directly
triggered the sync saga. As a result, the redux store was never updated
with the information that a sync process had begun. To solve the
problem, a new saga has been added that triggers the 'begin sync data'
action, which, in turn, triggers the sync saga.