Add update-metric saga
When updating a metric, its `lastModified` field gets set to the current
time. Since we cannot get the current time from the reducer, the
`REQUEST_UPDATE_METRIC` action triggers the updateMetric saga, which, in
turn, triggers the `SUCCESS_UPDATE_METRIC` action with the current time
as a property.

This has the further benefit that we can check for null-valued fields
and reject the update if any any are present.