Implement zoom animation saga for Charts
To animate zooming the data plots, this saga reacts to charts/REQUEST_ZOOM and updates the view in small steps.

If an animation is already in progress (i.e. the `animation` property of the Redux-store chart object in question is not null), the finishTime and targetZoomFactor of the running animation is taken into account when initiating the new animation.