Implement ScrollBar component for charts
The scroll bar element is draggable using react-draggable.
It uses react-measure to get the dimensions of the container.
I was unable to test the drag functionality because enzyme is making
simulation of drag events difficult.

For some reason, after dragging the scroll bar, it jumps to a point
roughly 1/5 of the dragged distance.
I haven't figured out where that factor is coming from.