Implement axis tick generators
For the y axis, the ticks are quite straight-forward;
five values from the range of valid values are selected.
For the x axis, there are a number of rules
to select appropriate points in time to annotate.

Since I'm using lodash/fp a lot here,
I've also added libdefs for fp.