Change Chart state shape to use msPerPx
Since the size of a chart may change over time,
we want to calculate axis ticks and such
based on the distance between each point,
in pixels.
This is achieved by replacing the `zoomFactor` variable
with the `msPerPx` variable.
Additionally, to prevent scrolling past the edges of the data,
the chart state now contains a tuple representing
the range of dates in the given metrics.